refactor: extract Google Photos OAuth state
